Agents base class
Create an agent base class to standardize the creation of agents.
